backend/requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg[binary,pool]==3.2.3
python-dotenv==1.0.1
passlib==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
pydantic[email]==2.9.2
